refactor : 메인 다이얼 개선
메인 다이얼의 버그를 개선함.
메인 다이얼의 성능을 개선함.

+++ app/src/main/java/com/example/plandial/SpinnableImageView.java
@@ -10,10 +10,10 @@ import android.view.animation.RotateAnimation;
 import androidx.annotation.Nullable;
 
 public class SpinnableImageView extends androidx.appcompat.widget.AppCompatImageView {
-    protected static final double INITIAL_DEGREE = 180;
+    protected static final float INITIAL_DEGREE = 180;
 
-    protected double currentDegree = 0;
-    protected Pair<Double, Double> lastTouchLocation = new Pair<>(0.0, 0.0);
+    protected float currentDegree = 0;
+    protected Pair<Float, Float> lastTouchLocation = new Pair<>(0.0f, 0.0f);
 
     public SpinnableImageView(Context context) {
         super(context);
@@ -33,19 +33,19 @@ public class SpinnableImageView extends androidx.appcompat.widget.AppCompatImage
     @Override
     public boolean onTouchEvent(final MotionEvent event) {
         // 그림의 중심을 원점으로 하여 터치된 위치의 벡터를 구함.
-        double x = event.getX() - (double) this.getWidth() / 2;
-        double y = (double) this.getHeight() / 2 - event.getY();
+        float x = event.getX() - (float) this.getWidth() / 2;
+        float y = (float) this.getHeight() / 2 - event.getY();
 
         if (event.getAction() == MotionEvent.ACTION_MOVE) {
             // 벡터의 내적을 구해 각도를 판정함.
-            double dotResult = lastTouchLocation.first * x + lastTouchLocation.second * y;
-            double theta = Math.toDegrees(Math.acos(dotResult / (Math.sqrt(x * x + y * y) * Math.sqrt(lastTouchLocation.first * lastTouchLocation.first + lastTouchLocation.second * lastTouchLocation.second))));
+            float dotResult = lastTouchLocation.first * x + lastTouchLocation.second * y;
+            float theta = (float) Math.toDegrees(Math.acos(dotResult / (Math.sqrt(x * x + y * y) * Math.sqrt(lastTouchLocation.first * lastTouchLocation.first + lastTouchLocation.second * lastTouchLocation.second))));
 
             // 벡터의 외적을 구해 회전 방향을 판정함.
-            double crossResult = lastTouchLocation.first * y - lastTouchLocation.second * x;
+            float crossResult = lastTouchLocation.first * y - lastTouchLocation.second * x;
             int direction = (crossResult > 0) ? -1 : 1;
 
-            double nextDegree = this.currentDegree + direction * theta;
+            float nextDegree = this.currentDegree + direction * theta;
 
             // nextDegree를 [0, 360)의 값으로 만듦.
             if (nextDegree < 0) {
@@ -62,7 +62,7 @@ public class SpinnableImageView extends androidx.appcompat.widget.AppCompatImage
         return true;
     }
 
-    protected void rotate(final double fromDegree, final double toDegree, final long duration) {
+    protected void rotate(final float fromDegree, final float toDegree, final long duration) {
         final RotateAnimation rotateAnimation = new RotateAnimation(
                 (float) fromDegree,
                 (float) toDegree,
